structure refactoring, remove unnecessary code
